+++ mqtt/ultrasonic_sensor.py
@@ -1,0 +1,42 @@
+import RPi.GPIO as GPIO
+import time
+
+class UltrasonicSensor:
+
+    #set GPIO Pins
+    GPIO_TRIGGER = 27
+    GPIO_ECHO = 17
+    
+    #set GPIO direction (IN / OUT)
+    GPIO.setup(GPIO_TRIGGER, GPIO.OUT)
+    GPIO.setup(GPIO_ECHO, GPIO.IN)
+
+    def __init__(self):
+        GPIO.setmode(GPIO.BCM)
+
+    def distance(self):
+        # set Trigger to HIGH
+        GPIO.output(self.GPIO_TRIGGER, True)
+    
+        # set Trigger after 0.01ms to LOW
+        time.sleep(0.00001)
+        GPIO.output(self.GPIO_TRIGGER, False)
+    
+        StartTime = time.time()
+        StopTime = time.time()
+    
+        # save StartTime
+        while GPIO.input(self.GPIO_ECHO) == 0:
+            StartTime = time.time()
+    
+        # save time of arrival
+        while GPIO.input(self.GPIO_ECHO) == 1:
+            StopTime = time.time()
+    
+        # time difference between start and arrival
+        TimeElapsed = StopTime - StartTime
+        # multiply with the sonic speed (34300 cm/s)
+        # and divide by 2, because there and back
+        distance = (TimeElapsed * 34300) / 2
+    
+        return distance
